Update infinite-scrolling.mdx
I was a little confused by the wording and meaning of the first sentence, so I moved the word "that" so that it would make more sense to me. But please make sure that it's correct. I figure there was a typo and wasn't sure what you had wanted it to say, so I just took my best guess.